async code

+++ qmq-delay-server/src/main/java/qunar/tc/qmq/delay/store/log/AbstractDelaySegment.java
@@ -40,7 +40,7 @@ public abstract class AbstractDelaySegment<T> implements DelaySegment<T> {
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractDelaySegment.class);
 
     private final File file;
-    private final int segmentBaseOffset;
+    private final long segmentBaseOffset;
     private final AtomicLong wrotePosition = new AtomicLong(0);
     private final AtomicLong flushedPosition = new AtomicLong(0);
     private final AtomicBoolean needFlush = new AtomicBoolean(true);
@@ -52,7 +52,7 @@ public abstract class AbstractDelaySegment<T> implements DelaySegment<T> {
     AbstractDelaySegment(File file) throws IOException {
         this.file = file;
         this.fileName = file.getAbsolutePath();
-        this.segmentBaseOffset = Integer.parseInt(file.getName());
+        this.segmentBaseOffset = Long.parseLong(file.getName());
         boolean success = false;
         try {
             fileChannel = new RandomAccessFile(file, "rw").getChannel();
@@ -122,7 +122,7 @@ public abstract class AbstractDelaySegment<T> implements DelaySegment<T> {
     }
 
     @Override
-    public int getSegmentBaseOffset() {
+    public long getSegmentBaseOffset() {
         return segmentBaseOffset;
     }
